// src/club/enums/club-message.enum.ts
export enum ClubMessage {
    NOT_FOUND = 'Club not found',
    ALREADY_EXISTS = 'Club already exists',
    CREATED = 'Club created successfully',
    UPDATED = 'Club updated successfully',
    DELETED = 'Club deleted successfully',
    GET_ALL_CLUBS = 'Get all clubs successfully',
    GET_CLUB = 'Get club successfully',
}